requirements.txt
# Satellite data access
pystac-client>=0.7.0
planetary-computer>=1.0.0
rasterio>=1.3.0
rioxarray>=0.15.0

# Geospatial
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
fiona>=1.9.0

# ML / Image processing
numpy>=1.24.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3
matplotlib>=3.7.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Utilities
python-dateutil>=2.8.0
requests>=2.31.0
tqdm>=4.65.0
python-dotenv>=1.0.0
joblib>=1.3.0

# Logging
structlog>=23.1.0
